requirements.txt
dotenv
openai
google-genai